requirements.txt
fastapi
uvicorn[standard]
azure-cognitiveservices-speech
openai==0.28
jinja2
